requirements.txt
certifi==2019.3.9
cffi==1.13.0
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.1.1
future==0.18.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pipenv==2018.11.26
psycopg2==2.8.4
pycparser==2.19
python-telegram-bot==12.2.0
requests==2.7.0
selenium==3.141.0
six==1.12.0
telebot==0.0.3
tinydb==3.15.1
tornado==6.0.3
urllib3==1.25.6
virtualenv==16.7.7
virtualenv-clone==0.5.1
Werkzeug==0.16.0
